requirements.txt
pyserial==3.5
python-dotenv==0.20.0
SQLAlchemy==1.4.35
